// components/WistListProducts.tsx
"use client"
import Container from '@/components/Container';
import { Button } from '@/components/ui/button';
import { Product } from '@/sanity.types';
import useStore from '@/store'
import { Heart, X } from 'lucide-react';
import Link from 'next/link';
import { useState } from 'react';
import toast from 'react-hot-toast';
import Image from 'next/image';
import { urlFor } from '@/sanity/lib/image';
import PriceFormatter from '@/components/PriceFormatter';
import AddToCartButton from '@/components/AddToCartButton';
const WistListProducts = () => {
    const [visibleProducts, setVisibleProducts] = useState(7);
    const { favoriteProduct, removeFromFavorite, resetFavorite } = useStore();
    const loadMore = () => {
        setVisibleProducts((prev) => Math.min(prev + 5, favoriteProduct.length))
    }
    const handleResetWishlist = () => {
        const confirmReset = window.confirm("Bạn có chắc chắn muốn đặt lại danh sách yêu thích của mình không?");
        if (confirmReset) {
            resetFavorite();
            toast.success("Xoá danh sách yêu thích thành công");
        }
    };

    return (
        <Container>
            {favoriteProduct?.length > 0 ? (
                <>
                    <div className='overflow-x-auto'>
                        <table className='w-full border-collapse'>
                            <thead className="border-b">
                                <tr className="bg-black/5">
                                    <th className="p-2 text-left">Hình ảnh</th>
                                    <th className="p-2 text-left hidden md:table-cell">Danh mục</th>
                                    <th className="p-2 text-left hidden md:table-cell">Phân loại</th>
                                    <th className="p-2 text-left hidden md:table-cell">Trạng thái</th>
                                    <th className="p-2 text-left">Giá tiền</th>
                                    <th className="p-2 text-center md:text-left">Thao tác</th>
                                </tr>
                            </thead>
                            <tbody>
                                {favoriteProduct
                                    ?.slice(0, visibleProducts)
                                    ?.map((product: Product) => (
                                        <tr key={product?._id} className="border-b">
                                            <td className='px-2 py-4 flex items-center'>
                                                <X
                                                    onClick={() => {
                                                        removeFromFavorite(product?._id);
                                                        toast.success("Product removed from wistlist")
                                                    }}
                                                    size={18}
                                                    className=' hover:text-red-600 hover:cursor-pointer hoverEffect'
                                                />
                                                {product?.images && (
                                                    <Link href={`/product/${product?.slug?.current}`}
                                                        className='border rounded-md group hidden md:inline-flex'
                                                    >
                                                        <Image
                                                            src={urlFor(product?.images[0]).url()}
                                                            alt={"product image"}
                                                            width={80}
                                                            height={80}
                                                            className="rounded-md group-hover:scale-105 hoverEffect h-20 w-20 object-contain"
                                                        />
                                                    </Link>
                                                )}
                                                <p className='line-clamp-1'>{product?.name}</p>
                                            </td>
                                            <td className="p-2 capitalize hidden md:table-cell">
                                                {product?.categories && (
                                                    <p className="uppercase line-clamp-1 text-xs font-medium">
                                                        {product.categories.map((cat) => cat).join(", ")}
                                                    </p>
                                                )}
                                            </td>
                                            <td className='p-2 capitalize hidden md:table-cell'>{product?.variant}</td>
                                            <td
                                                className={`p-2 w-24 ${(product?.stock as number) > 0
                                                    ? "text-green-600"
                                                    : "text-red-600"
                                                    } font-medium text-sm hidden md:table-cell`}
                                            >
                                                {(product?.stock as number) > 0 ? "Còn hàng" : "Hết hàng"}
                                            </td>
                                            <td className='p-2'>
                                                <PriceFormatter amount={product?.price} />
                                            </td>
                                            <td className='p-2'>
                                                <AddToCartButton product={product} className='w-full' />
                                            </td>
                                        </tr>
                                    ))}
                            </tbody>
                        </table>
                    </div>
                    <div className="flex items-center gap-2">
                        {visibleProducts < favoriteProduct?.length && (
                            <div className="my-5">
                                <Button variant="outline" onClick={loadMore}>
                                    Load More
                                </Button>
                            </div>
                        )}
                        {visibleProducts > 10 && (
                            <div className="my-5">
                                <Button
                                    onClick={() => setVisibleProducts(10)}
                                    variant="outline"
                                >
                                    Load Less
                                </Button>
                            </div>
                        )}
                    </div>
                    {favoriteProduct?.length > 0 && (
                        <Button
                            onClick={handleResetWishlist}
                            className="mb-5 font-semibold"
                            variant="destructive"
                            size="lg"
                        >
                            Xoá tất cả yêu thích
                        </Button>
                    )}
                </>
            ) : (
                <div className="flex min-h-[400px] flex-col items-center justify-center space-y-5 text-center">
                    <div className="relative mb-4">
                        <div className="absolute -top-1 -right-1 h-4 w-4 animate-ping rounded-full bg-muted-foreground/20" />
                        <Heart
                            className="h-12 w-12 text-muted-foreground"
                            strokeWidth={1.5}
                        />
                    </div>
                    <div className="space-y-2">
                        <h2 className="text-2xl font-semibold tracking-tight">
                            Your wishlist is empty
                        </h2>
                        <p className="text-sm text-muted-foreground">
                            Items added to your wishlist will appear here
                        </p>
                    </div>

                    <Button asChild >
                        <Link href="/shop">Continue Shopping</Link>
                    </Button>
                </div>
            )}
        </Container>
    )
}

export default WistListProducts
